[REFACTOR][S-TIR] Minimize src/support/ by relocating s_tir-private headers
This PR consolidates src/support/ by moving s_tir-private helpers into
a new src/s_tir/support/ directory, merging generic_arena.h into arena.h,
extracting fp-limits constants into limits.h, and deleting unused code.

Main changes:

- Move table_printer.h, array.h (→ array_utils.h), nd_int_set.h, and
  parallel_for.h (made header-only) to src/s_tir/support/
- Merge generic_arena.h content into arena.h; delete generic_arena.h
- Extract kMax* fp-limit constants from scalars.h into new limits.h;
  delete scalars.{h,cc} (unused functions) and hexdump.h
- Rewrite layout_transformation.cc to use std::vector instead of std::list,
  removing the support::AsList/AsArray dependency
