+++ src/s_tir/schedule/utils.h
@@ -42,8 +42,8 @@
 #include "../../arith/pattern_match.h"
 #include "../../ir/attr_registry.h"
 #include "../../runtime/thread_storage_scope.h"
-#include "../../support/array.h"
-#include "../../support/nd_int_set.h"
+#include "../support/array_utils.h"
+#include "../support/nd_int_set.h"
 #include "./analysis.h"
 #include "./error.h"
 #include "./instruction_traits.h"
