+++ src/s_tir/support/parallel_for.h
@@ -18,13 +18,19 @@
  */
 
 /*!
- * \file parallel_for.cc
+ * \file parallel_for.h
  * \brief An implementation to run loop in parallel.
  */
-#include <tvm/runtime/logging.h>
-#include <tvm/support/parallel_for.h>
+#ifndef TVM_S_TIR_SUPPORT_PARALLEL_FOR_H_
+#define TVM_S_TIR_SUPPORT_PARALLEL_FOR_H_
 
+#include <tvm/runtime/base.h>
+#include <tvm/runtime/logging.h>
+
+#include <atomic>
+#include <functional>
 #include <future>
+#include <mutex>
 #include <thread>
 #include <utility>
 #include <vector>
@@ -32,7 +38,19 @@
 namespace tvm {
 namespace support {
 
-std::vector<std::vector<int>> rr_partitioner(int begin, int end, int step, int num_threads) {
+using PartitionerFuncType = std::function<std::vector<std::vector<int>>(int, int, int, int)>;
+
+/*!
+ * \brief A partitioner to split the task to each thread in Round-robin manner.
+ * \param begin The start index of this parallel loop(inclusive).
+ * \param end The end index of this parallel loop(exclusive).
+ * \param step The traversal step to the index.
+ * \param num_threads The number of threads(the number of tasks to be partitioned to).
+ * \return A list with `num_threads` elements, and each is a list of integers indicating the loop
+ * indexes for the corresponding thread to process.
+ */
+TVM_DLL inline std::vector<std::vector<int>> rr_partitioner(int begin, int end, int step,
+                                                            int num_threads) {
   int total_task_count = (end - begin) / step;
   TVM_FFI_ICHECK_GE(total_task_count, 0)
       << "Infinite loop condition with begin: " << begin << " end: " << end << " step: " << step;
@@ -47,8 +65,28 @@ std::vector<std::vector<int>> rr_partitioner(int begin, int end, int step, int n
   return ret;
 }
 
-void parallel_for(int begin, int end, const std::function<void(int)>& f, int step,
-                  const PartitionerFuncType partitioner) {
+/*!
+ * \brief A runtime api provided to run the task function in parallel.
+ * e.g. A for loop:
+ *   for (int i = 0; i < 10; i++) {
+ *     a[i] = i;
+ *   }
+ * should work the same as:
+ *   parallel_for(0, 10, [&a](int index) {
+ *     a[i] = i;
+ *   });
+ * \param begin The start index of this parallel loop(inclusive).
+ * \param end The end index of this parallel loop(exclusive).
+ * \param f The task function to be executed. Assert to take an int index as input with no output.
+ * \param step The traversal step to the index.
+ * \param partitioner A partition function to split tasks to different threads. Use Round-robin
+ * partitioner by default.
+ * \note 1. Currently do not support nested parallel_for; 2. The order of execution in each thread
+ * is not guaranteed, the for loop task should be thread independent and thread safe.
+ */
+TVM_DLL inline void parallel_for(int begin, int end, const std::function<void(int)>& f,
+                                 int step = 1,
+                                 const PartitionerFuncType partitioner = rr_partitioner) {
   static bool GLOBAL_PARALLEL_FOR_FLAG{false};
   static std::mutex M_GLOBAL_PARALLEL_FOR_FLAG;
   {
@@ -94,8 +132,26 @@ void parallel_for(int begin, int end, const std::function<void(int)>& f, int ste
   }
 }
 
-void parallel_for_dynamic(int begin, int end, int num_threads,
-                          const std::function<void(int thread_id, int task_id)>& f) {
+/*!
+ * \brief An API to launch fix amount of threads to run the specific functor in parallel.
+ * Different from `parallel_for`, the partition is determined dynamically on the fly,
+ * i.e. any time when a thread is idle, it fetches the next task to run.
+ * The behavior is similar to dynamic scheduling in OpenMP:
+ *
+ *   \#pragma omp parallel for schedule(dynamic) num_threads(num_threads)
+ *   for (int i = 0; i < 10; i++) {
+ *     a[i] = i;
+ *   }
+ *
+ * \param begin The start index of this parallel loop (inclusive).
+ * \param end The end index of this parallel loop (exclusive).
+ * \param num_threads The number of threads to be used.
+ * \param f The task function to be executed. Takes the thread index and the task index as
+ * input with no output.
+ * \note `step` support is left for future work.
+ */
+TVM_DLL inline void parallel_for_dynamic(int begin, int end, int num_threads,
+                                         const std::function<void(int thread_id, int task_id)>& f) {
   // Step 1. Sanity checks
   if (begin == end) {
     return;
@@ -143,3 +199,5 @@ void parallel_for_dynamic(int begin, int end, int num_threads,
 
 }  // namespace support
 }  // namespace tvm
+
+#endif  // TVM_S_TIR_SUPPORT_PARALLEL_FOR_H_
