+++ src/s_tir/transform/compact_buffer_region.cc
@@ -34,10 +34,10 @@
 #include <stack>
 
 #include "../../support/arena.h"
-#include "../../support/nd_int_set.h"
 #include "../../support/utils.h"
 #include "../../tirx/transform/ir_utils.h"
 #include "../schedule/utils.h"
+#include "../support/nd_int_set.h"
 
 namespace tvm {
 namespace s_tir {
